+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ### Features
 + 添加新组件 Picker，选择器
 + 添加新组件 DatetimePicker，时间选择器
-
++ 在index中添加了兼容requestAnimationFrame函数
 ## 0.0.32 (2019-09-02)
 
 ### Features